Consolidate blitz_decode_weights into prepare_weights
