+++ models/demos/deepseek_v3_b1/fused_ops/moe_routed_expert/op.py
@@ -29,7 +29,7 @@ from models.demos.deepseek_v3_b1.unified_kernel_descriptor import (
 )
 
 if TYPE_CHECKING:
-    from models.demos.deepseek_v3_b1.blitz_decode_weights import OverlappedTensor
+    from models.demos.deepseek_v3_b1.blitz_overlap_tensors import OverlappedTensor
 
 # Device roles for ReduceToOneB1
 MESH_LEAF = 0
