+++ models/demos/deepseek_v3_b1/tests/unit_tests/test_decoder_block.py
@@ -18,7 +18,6 @@ import ttnn
 from models.common.utility_functions import comp_pcc
 from models.demos.deepseek_v3.reference.modeling_deepseek import yarn_get_mscale
 from models.demos.deepseek_v3.tt.rope import get_cos_sin_matrix, get_rot_transformation_mat
-from models.demos.deepseek_v3_b1.blitz_decode_weights import BlitzDecodeWeights
 from models.demos.deepseek_v3_b1.fused_ops.attention_block.op import AttentionBlock
 from models.demos.deepseek_v3_b1.fused_ops.decoder_block.op import DecoderBlock
 from models.demos.deepseek_v3_b1.fused_ops.moe.op import MoeOp
@@ -66,7 +65,7 @@ def _decode_expert_upload_mode(expert_upload_mode: str) -> tuple[int, int | None
 
 
 # ============================================================================
-# Unified decoder block tensor setup (single BDW instance for all L1 weights)
+# Unified decoder block tensor setup
 # ============================================================================
 def create_decoder_block_tensors(
     submesh,
@@ -89,10 +88,10 @@ def create_decoder_block_tensors(
 
     Three modes of operation:
     - **preloaded_weights mode** (production): pass a DeepSeekV3MoELayerWeights from
-      load_moe_layer. Skips BDW allocation, weight processing, and golden tensors.
-    - **state_dict + is_moe=True** (MoE tests): allocates BDW, calls
+      load_moe_layer. Skips weight processing and golden tensors.
+    - **state_dict + is_moe=True** (MoE tests): calls
       prepare_moe_layer_weights, builds MoE golden tensors.
-    - **state_dict + is_moe=False** (dense tests): allocates BDW, calls
+    - **state_dict + is_moe=False** (dense tests): calls
       prepare_dense_layer_weights, builds dense golden tensors.
 
     Returns a dict with all attention + FFN + shared expert + reduce tensors.
@@ -257,22 +256,21 @@ def create_decoder_block_tensors(
         )
 
     # ══════════════════════════════════════════════════════════════════════════
-    # All weights via a single BDW instance or preloaded
+    # All weights via prepare_*_layer_weights or preloaded
     # ══════════════════════════════════════════════════════════════════════════
     if preloaded_weights is not None:
         layer = preloaded_weights
     else:
-        bdw = BlitzDecodeWeights(submesh)
         if is_moe:
             layer = prepare_moe_layer_weights(
-                bdw,
+                submesh,
                 state_dict,
                 layer_idx,
                 num_routed_experts=num_routed_experts,
                 move_to_device=True,
             )
         else:
-            layer = prepare_dense_layer_weights(bdw, state_dict, layer_idx, move_to_device=True)
+            layer = prepare_dense_layer_weights(submesh, state_dict, layer_idx, move_to_device=True)
 
     # ── FFN final output config (DRAM streaming matmul output grid) ──
     gate_proj_noc = ttnn.NOC.NOC_0
@@ -781,7 +779,7 @@ def create_decoder_block_tensors(
     routed_down = layer.routed_down_proj[0] if is_moe else layer.routed_down_proj
 
     result = {
-        # Attention weights (from prepare_*_layer_weights via BDW)
+        # Attention weights (from prepare_*_layer_weights)
         "gamma_overlapped": layer.attn_norm,
         "matmul_weights_overlapped": layer.q_a_proj,
         "rmsnorm2_gamma_overlapped": layer.q_norm,
